Struct actors::channel::ActorCell

pub struct ActorCell<Message, A: Actor<Message>> {
    // some fields omitted
}

A simplistic environment to run an actor in which can act as ActorRef.

Currently, it still uses one thread per actor.

Methods

impl<Message: Send + 'static, A: 'static + Actor<Message>> ActorCell<Message, A>

fn create(actor: A) -> Arc<ActorCell<Message, A>>

Create and ActorCell for the given actor.

fn stop_and_join(&self) -> A

Stops the actor cell and returns the latest actor state.